+++ registry/components/ui/kbd/index.ts
@@ -1,2 +1,3 @@
 export { default as Kbd } from "./kbd.svelte";
 export { default as KbdGroup } from "./kbd-group.svelte";
+export { matchKeys } from "./shortcut.ts";